# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: 群想 AI 群聊 API
  version: 1.0.0
  description: |
    以 backend/src 真实路由为准的接口契约。覆盖认证、短信、群组、消息、文件、
    TTS、社交互动、记忆、Agent、互动分析、监控、画像、人设与用户 API 配置等域。
    部分复杂查询端点（搜索、互动分析、监控指标等）的响应体为宽松对象，
    字段以实现为准，后续逐步收紧。
servers:
  - url: http://localhost:3002
    description: 本地后端
  - url: https://{YOUR_BACKEND_HOST}
    description: 生产后端（占位符，替换为真实 Render/自有域名）
    variables:
      YOUR_BACKEND_HOST:
        default: your-backend-host.example.com
...
security:
  - cookieAuth: []
  - csrfHeader: []
paths:
...
  /api/sms/send:
    post:
      tags: [Sms]
      operationId: sendSmsCode
      summary: 发送短信验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneInput'
      responses:
        '200':
          description: 发送结果（阿里云返回体透传）
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: 发送过于频繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: 短信服务未配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/sms/verify:
    post:
      tags: [Sms]
      operationId: verifySmsCode
      summary: 校验短信验证码（不注册）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsVerifyInput'
      responses:
        '200':
          description: 校验结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: 验证码尝试次数过多
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: 短信服务未配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ------------------------------------------------------------------ Groups
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: 登录端点下发的 httpOnly 会话 Cookie
    csrfHeader:
      type: apiKey
      in: header
      name: x-csrf-token
      description: |
        所有非安全方法（POST/PUT/PATCH/DELETE）必须携带，
        令牌通过 GET /api/csrf-token 获取，且须与 Cookie 中的
        CSRF 值一致（double-submit 模式）
...
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PhoneInput:
      type: object
      required: [phone]
      additionalProperties: false
      properties:
        phone:
          type: string
          pattern: '^1[3-9][0-9]{9}$'
...
    SmsVerifyInput:
      type: object
      required: [phone, code]
      additionalProperties: false
      properties:
        phone:
          type: string
          pattern: '^1[3-9][0-9]{9}$'
        code:
          type: string
          pattern: '^[0-9]{4,8}$'
...